move rss response processor to rss package
